+++ central/role/utils/utils_test.go
@@ -8,18 +8,6 @@ import (
 	"github.com/stretchr/testify/assert"
 )
 
-func TestFillAccessList(t *testing.T) {
-	testRole := &storage.Role{
-		GlobalAccess: storage.Access_READ_WRITE_ACCESS,
-		ResourceToAccess: map[string]storage.Access{
-			"Alert": storage.Access_READ_ACCESS,
-		},
-	}
-
-	FillAccessList(testRole)
-	assert.Equal(t, testRole.GetResourceToAccess()["Alert"], storage.Access_READ_WRITE_ACCESS)
-}
-
 func TestValidateSimpleAccessScope(t *testing.T) {
 	mockGoodID := EnsureValidAccessScopeID("42")
 	mockBadID := "42"