+++ migrator/runner/all.go
@@ -54,4 +54,5 @@ import (
 	_ "github.com/stackrox/rox/migrator/migrations/m_65_to_m_66_policy_bug_fixes"
 	_ "github.com/stackrox/rox/migrator/migrations/m_66_to_m_67_missing_policy_migrations"
 	_ "github.com/stackrox/rox/migrator/migrations/m_67_to_m_68_exclude_pdcsi_from_mount_propagation"
+	_ "github.com/stackrox/rox/migrator/migrations/m_68_to_m_69_update_global_access_roles"
 )